@@ -21,13 +21,13 @@ class CommonMarkMarkdown implements MarkdownInterface {
 	protected $converter;
 
 	/**
-	 * @var array
+	 * @var array<string, mixed>
 	 */
 	protected $_defaultConfig = [
 	];
 
 	/**
-	 * @param array $config
+	 * @param array<string, mixed> $config
 	 */
 	public function __construct(array $config = []) {
 		$this->setConfig($config);
@@ -35,7 +35,7 @@ class CommonMarkMarkdown implements MarkdownInterface {
 
 	/**
 	 * @param string $text
-	 * @param array $options
+	 * @param array<string, mixed> $options
 	 *
 	 * @return string
 	 */
@@ -46,7 +46,7 @@ class CommonMarkMarkdown implements MarkdownInterface {
 	}
 
 	/**
-	 * @param array $options
+	 * @param array<string, mixed> $options
 	 *
 	 * @return \League\CommonMark\MarkdownConverterInterface
 	 */
@@ -59,7 +59,7 @@ class CommonMarkMarkdown implements MarkdownInterface {
 	}
 
 	/**
-	 * @param array $options
+	 * @param array<string, mixed> $options
 	 *
 	 * @return \League\CommonMark\MarkdownConverterInterface
 	 */
